Read software revision information from file
Add a file containing the software revision to use for images signed by
this tool. We put this in a file over using an environment variable or
hard-coding it so the revision can be read by other external tooling.

Set the default to 1, this is lowest revision allowed by ROM on newer
K3 devices.

Signed-off-by: Andrew Davis <afd@ti.com>